configuration/pyproject.toml
[project]
name = "configuration"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "jinja2>=3.1.6",
    "jinja2-time>=0.2.0",
    "json5>=0.12.0",
    "pyvis>=0.3.2",
    "pyyaml>=6.0.2",
]
[project.scripts]
buildconfig = "buildtools/build.py"
